connection-manager: moved ConnectionState to its own module
