+++ ouroboros-network-framework/ouroboros-network-framework.cabal
@@ -31,6 +31,7 @@ library
     Ouroboros.Network.ConnectionId
     Ouroboros.Network.ConnectionManager.Core
     Ouroboros.Network.ConnectionManager.InformationChannel
+    Ouroboros.Network.ConnectionManager.State
     Ouroboros.Network.ConnectionManager.Types
     Ouroboros.Network.Context
     Ouroboros.Network.Driver
